add webuploader implement the wifi upload function

+++ PasswordMini/PMMainViewController.m
@@ -71,11 +71,11 @@
     
     UILabel *label = (UILabel *)[view viewWithTag:1];
     if ([kind isEqualToString:UICollectionElementKindSectionHeader]){
-        label.text = [NSString stringWithFormat:@"这是header:%d",indexPath.section];
+        label.text = [NSString stringWithFormat:@"这是header:%ld",(long)indexPath.section];
     }
     else if ([kind isEqualToString:UICollectionElementKindSectionFooter]){
         view.backgroundColor = [UIColor lightGrayColor];
-        label.text = [NSString stringWithFormat:@"这是footer:%d",indexPath.section];
+        label.text = [NSString stringWithFormat:@"这是footer:%ld",(long)indexPath.section];
     }
     return view;
 }
